Fix issue prventing PRBS20 from ending
Changed the size_t in loops to the appropriate uint8, uint16, or uint32

+++ bare_bones/bare_bones.ino
@@ -29,7 +29,7 @@ void setupPRBSOutPin(uint8_t pin) {
 void PRBS7(uint8_t start)
 {
   uint8_t a = start;
-  for (size_t i = 0; i < 127; ++i)
+  for (uint8_t i = 0; i < 127; ++i)
   {
     bool newbit = ((a >> 6) ^ (a >> 5)) & 1;
     a = ((a << 1) | newbit) & 0x7f;
@@ -49,7 +49,7 @@ void PRBS7(uint8_t start)
 void PRBS13(uint16_t start)
 {
   uint16_t a = start;
-  for (size_t i = 0; i < 8191; ++i)
+  for (uint16_t i = 0; i < 8191; ++i)
   {
     bool newbit = ((a >> 12) ^ (a >> 11) ^ (a >> 1) ^ (a >> 0)) & 1;
     a = ((a << 1) | newbit) & 0x1fff;
@@ -68,7 +68,7 @@ void PRBS13(uint16_t start)
 void PRBS20(uint32_t start)
 {
   uint32_t a = start;
-  for (size_t i = 0; i < 1048575; ++i) 
+  for (uint32_t i = 0; i < 1048575; ++i) 
   {
     bool newbit = ((a >> 19) ^ (a >> 2)) & 1;
     a = ((a << 1) | newbit) & 0xfffff;
